SageMath version 8.8.beta7, Release Date: 2019-05-30

+++ src/sage/matrix/misc.pyx
@@ -420,7 +420,7 @@ def matrix_rational_echelon_form_multimodular(Matrix self, height_guess=None, pr
             verbose("Not checking validity of result (since proof=False).", level=2, caller_name="multimod echelon")
             break
         d   = E.denominator()
-        hdE = long((d*E).height())
+        hdE = int((d*E).height())
         if hdE * self.ncols() * height < prod:
             verbose("Validity of result checked.", level=2, caller_name="multimod echelon")
             break
